getDefaultProfilePicture is pulled out of Profile
