+++ Mountain-UI-Companion/Account/Profile/ProfileModel/Profile.swift
@@ -30,18 +30,7 @@ class Profile
         self.profilePictureURL = profilePictureURL
     }
     
-    // TODO: Move out, should be UI independent
-    public func getDefaultProfilePicture(fontSize: CGFloat) -> UILabel {
-        let initialsLabel = UILabel()
-        
-        initialsLabel.text = name.initials
-        initialsLabel.textColor = .label
-        initialsLabel.font = UIFont.systemFont(ofSize: fontSize, weight: .medium)
-        
-        return initialsLabel
-    }
-    
-    public static func createProfile(id: String, firstName: String, lastName: String, email: String, profilePictureURL: String? = nil, completion: @escaping (Profile) -> Void) {
+    public static func createProfile(uuid: String, firstName: String, lastName: String, email: String, profilePictureURL: String? = nil, completion: @escaping (Profile) -> Void) {
         guard let profilePictureURL = URL(string: profilePictureURL ?? "") else {
             completion(Profile(id: id, firstName: firstName, lastName: lastName, email: email))
             return
